Clean up C code, remove nested functions
Allows building in non-gcc compilers (read: clang)
