+++ cloudfuse.c
@@ -16,25 +16,8 @@
 #include <stddef.h>
 #include "cloudfsapi.h"
 #include "config.h"
+#include "cloudfuse.h"
 
-#define OPTION_SIZE 1024
-static int cache_timeout;
-
-typedef struct dir_cache
-{
-  char *path;
-  dir_entry *entries;
-  time_t cached;
-  struct dir_cache *next, *prev;
-} dir_cache;
-static dir_cache *dcache;
-static pthread_mutex_t dmut;
-
-typedef struct
-{
-  int fd;
-  int flags;
-} openfile;
 
 static void dir_for(const char *path, char *dir)
 {
@@ -398,39 +381,25 @@ char *get_home_dir()
   return "~";
 }
 
+int parse_option(void *data, const char *arg, int key, struct fuse_args *outargs)
+{
+  if (sscanf(arg, " username = %[^\r\n ]", options.username) ||
+      sscanf(arg, " api_key = %[^\r\n ]", options.api_key) ||
+      sscanf(arg, " cache_timeout = %[^\r\n ]", options.cache_timeout) ||
+      sscanf(arg, " authurl = %[^\r\n ]", options.authurl) ||
+      sscanf(arg, " use_snet = %[^\r\n ]", options.use_snet))
+    return 0;
+  if (!strcmp(arg, "-f") || !strcmp(arg, "-d") || !strcmp(arg, "debug"))
+    cloudfs_debug(1);
+  return 1;
+}
+
 int main(int argc, char **argv)
 {
   char settings_filename[MAX_PATH_SIZE] = "";
   FILE *settings;
   struct fuse_args args = FUSE_ARGS_INIT(argc, argv);
-
-  struct options {
-      char username[OPTION_SIZE];
-      char api_key[OPTION_SIZE];
-      char cache_timeout[OPTION_SIZE];
-      char authurl[OPTION_SIZE];
-      char use_snet[OPTION_SIZE];
-  } options = {
-      .username = "",
-      .api_key = "",
-      .cache_timeout = "600",
-      .authurl = "https://auth.api.rackspacecloud.com/v1.0",
-      .use_snet = "false",
-  };
-
-  int parse_option(void *data, const char *arg, int key, struct fuse_args *outargs)
-  {
-    if (sscanf(arg, " username = %[^\r\n ]", options.username) ||
-        sscanf(arg, " api_key = %[^\r\n ]", options.api_key) ||
-        sscanf(arg, " cache_timeout = %[^\r\n ]", options.cache_timeout) ||
-        sscanf(arg, " authurl = %[^\r\n ]", options.authurl) ||
-        sscanf(arg, " use_snet = %[^\r\n ]", options.use_snet))
-      return 0;
-    if (!strcmp(arg, "-f") || !strcmp(arg, "-d") || !strcmp(arg, "debug"))
-      cloudfs_debug(1);
-    return 1;
-  }
-
+ 
   fuse_opt_parse(&args, &options, NULL, parse_option);
 
   snprintf(settings_filename, sizeof(settings_filename), "%s/.cloudfuse", get_home_dir());
